+++ core/class/scenario.class.php
@@ -904,7 +904,7 @@ class scenario {
 				return true;
 			}
 		}
-		if (shell_exec('ps ax | grep -ie "scenario_id=' . $this->getId() . ' force" | wc -l') > 0) {
+		if (shell_exec('ps ax | grep -ie "scenario_id=' . $this->getId() . ' " | grep -v grep | wc -l') > 0) {
 			return true;
 		}
 		return false;
@@ -912,21 +912,23 @@ class scenario {
 
 	public function stop() {
 		if ($this->running()) {
-			$kill = posix_kill($this->getPID(), 15);
-			$retry = 0;
-			while (!$kill && $retry < 5) {
-				sleep(1);
-				$kill = posix_kill($this->getPID(), 9);
-				$retry++;
-			}
-			$retry = 0;
-			while ($this->running() && $retry < 5) {
-				sleep(1);
-				exec('kill -9 ' . $this->getPID());
-				$retry++;
+			if ($this->getPID() > 0) {
+				$kill = posix_kill($this->getPID(), 15);
+				$retry = 0;
+				while (!$kill && $retry < 5) {
+					sleep(1);
+					$kill = posix_kill($this->getPID(), 9);
+					$retry++;
+				}
+				$retry = 0;
+				while ($this->running() && $retry < 5) {
+					sleep(1);
+					exec('kill -9 ' . $this->getPID());
+					$retry++;
+				}
 			}
 			if ($this->running()) {
-				exec("ps aux | grep -ie 'scenario_id=" . $this->getId() . " force' | awk '{print $2}' | xargs kill -9 > /dev/null 2>&1");
+				exec("ps aux | grep -ie 'scenario_id=" . $this->getId() . " ' | grep -v grep | awk '{print $2}' | xargs kill -9 > /dev/null 2>&1");
 			}
 			if ($this->running()) {
 				throw new Exception(__('Impossible d\'arrêter le scénario : ', __FILE__) . $this->getHumanName() . __('. PID : ', __FILE__) . $this->getPID());
